Remove implementation for fromXML/toXML
